requirements.txt
# Core dependencies
fastapi>=0.95.1
uvicorn>=0.22.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Database drivers
pymysql>=1.0.3
aiomysql>=0.1.1
pyodbc>=4.0.39

# HTTP client
httpx>=0.24.0

# Redis for stream processing
redis>=4.5.5
# aioredis=1.3.1


# More
aioodbc>=0.4.0

# Testing
# pytest>=7.3.1
# pytest-asyncio>=0.21.0

# Utilities
# python-multipart>=0.0.6
# email-validator>=2.0.0
# web-vitals>=3.0.0

# Documentation
# markupsafe>=2.1.2
# jinja2>=3.1.2
# starlette>=0.27.0
